Implement most simplest AND ANN
This implements the most simplest logical AND neural network I can imagine.

How does it works? For each input vector we calculate the total
stimulation for the output neuron. We let pass the input values in case
it is at least that high as the input threshold (means the neuron was
stimulated). Otherwise we set this to 0.

The same logic applies to the output neuron. In case the total input
(stimulation) reaches a certain level the neuron fires.